Poprawienie refaktoryzacji w modyfikowanych klasach oraz dodanie JavaDoca do JsonToolsController (a także obsługę błędów do funkcji filter)
