+++ src/main/java/pl/put/poznan/tools/app/JsonToolsApplication.java
@@ -1,4 +1,4 @@
-package pl.put.poznan.transformer.app;
+package pl.put.poznan.tools.app;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
