+++ src/main/java/pl/put/poznan/tools/logic/JsonInterpreter.java
@@ -1,4 +1,4 @@
-package pl.put.poznan.transformer.logic;
+package pl.put.poznan.tools.logic;
 
 /**
  * This is just an example to show that the logic should be outside the REST service.
